faiss-samples/pyproject.toml
[tool.poetry]
name = "faiss-samples"
version = "0.1.0"
description = ""
authors = ["Gavin Noronha <gavinln@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
ipython = "^8.24.0"
faiss-cpu = {version = "^1.8.0", platform = "linux"}
packaging = "^24.0"
fire = "^0.6.0"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
